backport Twig extension enhancement to 1.x
