+++ config/twig.xml
@@ -13,7 +13,11 @@
 
         <service id="Vich\UploaderBundle\Twig\Extension\UploaderExtension" public="false">
             <tag name="twig.extension" />
+        </service>
+
+        <service id="Vich\UploaderBundle\Twig\Extension\UploaderExtensionRuntime" public="false">
             <argument type="service" id="Vich\UploaderBundle\Templating\Helper\UploaderHelper" />
+            <tag name="twig.runtime" />
         </service>
 
     </services>
